pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "cvefinder-cli"
dynamic = ["version"]
description = "Official CLI for CVEFinder.io API"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "CVEFinder.io", email = "support@cvefinder.io"}
]
keywords = ["security", "cve", "vulnerability", "scanning", "cvefinder"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.7"
dependencies = [
    "aiohttp>=3.9.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tabulate>=0.9.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://cvefinder.io"
Documentation = "https://docs.cvefinder.io"
Repository = "https://github.com/cvefinder/cli"
Issues = "https://github.com/cvefinder/cli/issues"

[project.scripts]
cvefinder = "cvefinder.cli:main"

[tool.setuptools]
packages = ["cvefinder"]

[tool.setuptools_scm]
